Refactor archiver packages
- Move classes into corresponding packages
