+++ src/main/java/org/phoebus/channelfinder/Application.java
@@ -21,8 +21,8 @@ import java.util.List;
 import java.util.ServiceLoader;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import org.phoebus.channelfinder.configuration.ChannelProcessor;
 import org.phoebus.channelfinder.configuration.PopulateDBConfiguration;
-import org.phoebus.channelfinder.processors.ChannelProcessor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
