+++ src/main/java/org/phoebus/channelfinder/rest/controller/ChannelProcessorManager.java
@@ -8,10 +8,10 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import org.phoebus.channelfinder.entity.Channel;
 import org.phoebus.channelfinder.entity.Scroll;
-import org.phoebus.channelfinder.processors.ChannelProcessorInfo;
 import org.phoebus.channelfinder.rest.api.IChannelScroll;
 import org.phoebus.channelfinder.service.AuthorizationService;
 import org.phoebus.channelfinder.service.ChannelProcessorService;
+import org.phoebus.channelfinder.service.model.archiver.ChannelProcessorInfo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
