+++ src/main/java/org/phoebus/channelfinder/service/ChannelProcessorService.java
@@ -8,9 +8,9 @@ import java.util.Spliterator;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
+import org.phoebus.channelfinder.configuration.ChannelProcessor;
 import org.phoebus.channelfinder.entity.Channel;
-import org.phoebus.channelfinder.processors.ChannelProcessor;
-import org.phoebus.channelfinder.processors.ChannelProcessorInfo;
+import org.phoebus.channelfinder.service.model.archiver.ChannelProcessorInfo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.task.TaskExecutor;
