+++ src/main/java/org/phoebus/channelfinder/service/model/archiver/aa/ArchiveAction.java
@@ -1,4 +1,4 @@
-package org.phoebus.channelfinder.processors.aa;
+package org.phoebus.channelfinder.service.model.archiver.aa;
 
 public enum ArchiveAction {
   ARCHIVE("/archivePV"),
