+++ src/main/resources/META-INF/services/org.phoebus.channelfinder.configuration.ChannelProcessor
@@ -1,0 +1,1 @@
+org.phoebus.channelfinder.configuration.AAChannelProcessor
