--- src/main/resources/META-INF/services/org.phoebus.channelfinder.processors.ChannelProcessor
@@ -1,1 +1,0 @@
-org.phoebus.channelfinder.processors.aa.AAChannelProcessor
