+++ src/test/java/org/phoebus/channelfinder/processors/aa/AAChannelProcessorTest.java
@@ -11,6 +11,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.phoebus.channelfinder.service.model.archiver.aa.ArchivePVOptions;
 
 class AAChannelProcessorTest {
